+++ src/components/StartScreen/StartScreen.jsx
@@ -2,12 +2,12 @@ import styled from 'styled-components/macro'
 import Header from '../Header/Header'
 import { ReactComponent as GetThingsDoneSVG } from '../../assets/get-things-done.svg'
 
-export default function StartScreen() {
+export default function StartScreen({ onStart }) {
   return (
     <StartScreenGrid>
-      <Header>{'Let’s get things done'}</Header>
+      <Header text="Let’s get things done" />
       <GetThingsDoneSVG />
-      <StartTimerButton>Start Timer</StartTimerButton>
+      <StartTimerButton onClick={onStart}>Start Timer</StartTimerButton>
     </StartScreenGrid>
   )
 }
@@ -19,5 +19,10 @@ const StartScreenGrid = styled.div`
 const StartTimerButton = styled.button`
   width: auto;
   height: 50px;
-  background: linear-gradient(125deg, #a4e3cc #56dfd1);
+  background: linear-gradient(125deg, #a4e3cc, #56dfd1);
+  border-radius: 8px;
+  border: none;
+  color: white;
+  font-size: 20px;
+  padding: 0;
 `
