+++ src/index.js
@@ -4,12 +4,15 @@ import App from './components/App/App'
 import reportWebVitals from './reportWebVitals'
 import GlobalStyle from './GlobalStyle'
 import GlobalFonts from './fonts/fonts'
+import { BrowserRouter as Router } from 'react-router-dom'
 
 ReactDOM.render(
   <React.StrictMode>
-    <GlobalStyle />
-    <GlobalFonts />
-    <App />
+    <Router>
+      <GlobalStyle />
+      <GlobalFonts />
+      <App />
+    </Router>
   </React.StrictMode>,
   document.getElementById('root')
 )
